requirements.txt
argcomplete
pyyaml